refactor: enum 클래스들 패키지 변경
- Bridge: bridge/model/domain -> bridge/constant
- BridgeSign: bridge/model/domain -> bridge/constant
- SuccessOrFail: bridge/model/domain -> bridge/constant
