+++ src/main/java/bridge/BridgeMaker.java
@@ -1,7 +1,7 @@
 package bridge;
 
+import bridge.constant.Bridge;
 import bridge.constant.Constant;
-import bridge.model.domain.Bridge;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
