+++ src/main/java/bridge/constant/BridgeSign.java
@@ -1,4 +1,4 @@
-package bridge.model.domain;
+package bridge.constant;
 
 public enum BridgeSign {
     MOVABLE("O"),
