+++ src/main/java/bridge/model/domain/BridgeGame.java
@@ -1,6 +1,7 @@
 package bridge.model.domain;
 
 import bridge.constant.Constant;
+import bridge.constant.SuccessOrFail;
 import bridge.model.dto.GameResultDto;
 import bridge.model.dto.MoveResultDto;
 import java.util.List;
