+++ src/main/java/bridge/view/InputView.java
@@ -1,9 +1,9 @@
 package bridge.view;
 
+import bridge.constant.Bridge;
 import bridge.constant.Constant;
 import bridge.constant.ErrorMessage;
 import bridge.constant.Message;
-import bridge.model.domain.Bridge;
 import camp.nextstep.edu.missionutils.Console;
 import java.util.function.Consumer;
 import java.util.regex.Pattern;
